% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect.R
\name{auto_detect_covariates}
\alias{auto_detect_covariates}
\title{Auto-detect Covariates}
\usage{
auto_detect_covariates(data, treatment_var, exclude = NULL)
}
\arguments{
\item{data}{Data frame}

\item{treatment_var}{Character. Treatment variable name}

\item{exclude}{Character vector. Variables to exclude}
}
\value{
Character vector. Covariate names
}
\description{
Automatically identify which variables should be balanced
}
\keyword{internal}
