% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect.R
\name{auto_detect_treatment}
\alias{auto_detect_treatment}
\title{Auto-detect Treatment Variable}
\usage{
auto_detect_treatment(data, treatment_var = NULL)
}
\arguments{
\item{data}{Data frame}

\item{treatment_var}{Character. Manual specification (overrides auto-detection)}
}
\value{
Character. Name of treatment variable
}
\description{
Automatically identify the treatment/group variable in a dataset
}
\keyword{internal}
