% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect.R
\name{auto_detect_treatment_level}
\alias{auto_detect_treatment_level}
\title{Auto-detect Treatment Level}
\usage{
auto_detect_treatment_level(data, treatment_var, treatment_level = NULL)
}
\arguments{
\item{data}{Data frame}

\item{treatment_var}{Character. Treatment variable name}

\item{treatment_level}{Character/numeric. Manual specification}
}
\value{
Character/numeric. Treatment level
}
\description{
Automatically identify which level is the "treatment" group
}
\keyword{internal}
