% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boss_delta_cpp}
\alias{boss_delta_cpp}
\title{Change in balance from a 1-exchange}
\usage{
boss_delta_cpp(NT, NC, off, code_out, code_in, normalization = 0L)
}
\arguments{
\item{off}{0-based start position of each covariate in the flat vector}

\item{code_out}{bin code of the leaving unit, per covariate (1-based)}

\item{code_in}{bin code of the entering unit}
}
\description{
Change in balance from a 1-exchange
}
\keyword{internal}
