% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boss_objective_cpp}
\alias{boss_objective_cpp}
\title{Full balance from the bin counts}
\usage{
boss_objective_cpp(NT, NC, normalization = 0L)
}
\description{
Needed only at the start and after a restart.
}
\keyword{internal}
