% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{build_2way_table}
\alias{build_2way_table}
\title{Build 2-way comparison table}
\usage{
build_2way_table(
  data_before,
  data_mobaco,
  treatment_var,
  treatment_level,
  covariates,
  level_structure,
  solution_label
)
}
\description{
Build 2-way comparison table
}
\keyword{internal}
