% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{build_3way_table}
\alias{build_3way_table}
\title{Build 3-way comparison table}
\usage{
build_3way_table(
  data_before,
  data_mobaco,
  data_psm,
  treatment_var,
  treatment_level,
  covariates,
  level_structure,
  solution_label,
  psm_caliper = 0.2
)
}
\description{
Build 3-way comparison table
}
\keyword{internal}
