% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{build_level_structure}
\alias{build_level_structure}
\title{Build Level Structure}
\usage{
build_level_structure(data, treatment_var, treatment_level, covariates)
}
\description{
Build Level Structure
}
\keyword{internal}
