% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculate_psm_point}
\alias{calculate_psm_point}
\title{Calculate PSM point for Pareto plot}
\usage{
calculate_psm_point(mobaco_result)
}
\description{
Calculate PSM point for Pareto plot
}
\keyword{internal}
