% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{comparison_table}
\alias{comparison_table}
\title{Create Comparison Table}
\usage{
comparison_table(
  mobaco_result,
  solution = "min_balance",
  include_psm = NULL,
  psm_caliper = 0.2
)
}
\arguments{
\item{mobaco_result}{A mobaco object}

\item{solution}{Character. Which solution:
"min_balance" (default), "knee", or "max_coverage"}

\item{include_psm}{Logical. Include PSM? (default: TRUE if PSM was run)}

\item{psm_caliper}{Numeric. PSM caliper (default: 0.2)}
}
\description{
Create Comparison Table
}
