% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract_solution}
\alias{extract_solution}
\title{Extract Selected Solution Data}
\usage{
extract_solution(object, solution = "min_balance")
}
\arguments{
\item{object}{A mobaco object}

\item{solution}{Character. Which solution:
"min_balance" (default), "knee", or "max_coverage"}
}
\value{
A data frame with selected patients
}
\description{
Extract Selected Solution Data
}
