% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{get_table_cols}
\alias{get_table_cols}
\title{Get Table Columns with Fixed Structure}
\usage{
get_table_cols(
  data,
  treatment_var,
  treatment_level,
  covariates,
  level_structure
)
}
\description{
Get Table Columns with Fixed Structure
}
\keyword{internal}
