% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobaco-package.R
\docType{package}
\name{MOBACO-package}
\alias{MOBACO}
\alias{MOBACO-package}
\title{MOBACO: Multi-Objective Balance-Coverage Optimization for Observational Studies}
\description{
Constructs balanced cohorts for observational studies by jointly optimizing covariate balance and sample coverage with the NSGA-II multi-objective genetic algorithm. Rather than returning a single adjusted cohort, the method approximates the Pareto front of non-dominated subsets, making the balance-coverage trade-off explicit at the design stage so that the analyst can select an operating point before outcome information is used. Covariate imbalance is measured with the BOSS-B metric of Nikolaev et al. (2013) \doi{10.1287/opre.1120.1118}. Includes tools for comparison with propensity score matching, visualization of the Pareto front, and an optional Shiny interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sysbio-bioinf/MOBACO}
  \item Report bugs at \url{https://github.com/sysbio-bioinf/MOBACO/issues}
}

}
\author{
\strong{Maintainer}: Metehan Oruc \email{metehan.oruc@uni-ulm.de} (\href{https://orcid.org/0009-0007-1043-3626}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Metehan Oruc \email{metehan.oruc@uni-ulm.de} (\href{https://orcid.org/0009-0007-1043-3626}{ORCID}) [copyright holder]
  \item Johann M. Kraus (\href{https://orcid.org/0000-0002-9534-6295}{ORCID}) [copyright holder]
  \item Kathrin Stucke-Straub (\href{https://orcid.org/0009-0008-0697-534X}{ORCID}) [copyright holder]
  \item Hans A. Kestler (\href{https://orcid.org/0000-0002-4759-5254}{ORCID}) [copyright holder]
}

}
\keyword{internal}
