% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{mobaco_app}
\alias{mobaco_app}
\title{Launch MOBACO Shiny App}
\usage{
mobaco_app(launch.browser = TRUE, port = NULL)
}
\arguments{
\item{launch.browser}{Logical. Open in browser? (default: TRUE)}

\item{port}{Integer. Port number (default: NULL = auto-assign)}
}
\description{
Interactive web interface for Multi-Objective Balance and Coverage
Optimization. No programming required.
}
\examples{
\dontrun{
# Launch MOBACO Shiny App
mobaco_app()
}

}
