% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for MOBACO}
\usage{
prepare_data(
  data,
  treatment_var,
  treatment_level,
  covariates = NULL,
  n_bins = 5,
  verbose = TRUE
)
}
\description{
Prepare Data for MOBACO
}
\keyword{internal}
