% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mobaco}
\alias{print.mobaco}
\title{Print Method for MOBACO}
\usage{
\method{print}{mobaco}(x, ...)
}
\arguments{
\item{x}{A mobaco object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for MOBACO
}
