% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nsga2.R
\name{run_nsga2}
\alias{run_nsga2}
\title{Run NSGA-II Optimization}
\usage{
run_nsga2(
  fitness_fn,
  n_individuals,
  popsize,
  generations,
  crossover_prob = 0.8,
  mutation_prob = 0.1,
  crossover_dist = 10,
  mutation_dist = 10,
  verbose
)
}
\description{
Run NSGA-II Optimization
}
\keyword{internal}
