% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{smd_binary}
\alias{smd_binary}
\title{Calculate SMD for binary/categorical variables}
\usage{
smd_binary(x1, x0, level)
}
\description{
Calculate SMD for binary/categorical variables
}
\keyword{internal}
