% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mobaco}
\alias{summary.mobaco}
\title{Summary Method for MOBACO}
\usage{
\method{summary}{mobaco}(object, ...)
}
\arguments{
\item{object}{A mobaco object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for MOBACO
}
