% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect.R
\name{validate_data}
\alias{validate_data}
\title{Smart Data Validation}
\usage{
validate_data(data, treatment_var, treatment_level, covariates)
}
\description{
Validate and prepare data with helpful error messages
}
\keyword{internal}
